Stop using '__asm__' and 'asm' keywords
There was no consistency across the source tree about which keyword
introduces inline assembly. '__asm__' is less universally supported than
'asm' or '__asm' - for example, it is not permitted by IAR.

'asm' is more formally supported in C++. Though many C compilers accept it,
some will reject it if strict conformance is specified. As a result, the
most compatible choice is '__asm'.