Remove double typedefs
'alarm_pool_t' was typedeffed to 'struct alarm_pool' twice: once in
"pico/time.h" and once in time.c. Even though the struct was an incomplete
type in the header file, this double typedef isn't strictly permitted in C
(though it is in C++ as long as they're identical) and IAR was whinging
about it.

A similar problem applied to rom_table_lookup_fn in "pico/bootrom.h" and
bootrom.c, although in this case the type is a function pointer and is
identical in both places.