Replace data-emitting directives with macros
Macro names match GAS directive name except for replacing the leading '.'
with an '_'. This retains an element of familiarity for people used to GAS.
To remain valid macro syntax for all toolchains, leading whitespace is added
if there wasn't any already - 4 spaces are used for consistency with the
most common indentation previously used.