Replace .section, .data and .text directives with _section macro
Since the _section macro can't derive the syntax of the section name for
both GAS and IAR from a single parameter, it needs the caller to pass both,
This requirement cascades to several other macros that call _section, made
worse by various compile-time switches that determine whether the sections
end up in RAM or flash.

This is really ugly and not very maintainable, but the alternative is to
rename all the sections to lose the '.' characters to enable IAR to lose the
backtick quote requirement and to do name string manipulation like GAS does.