Rename section so it doesn't start with a digit (macro argument compat)
The problem is that section names end up being passed to the MA() macro, and
in its IAR variant, this leaves the section name unchanged. But when the
preprocessor looks at this to see if it can be further expanded, it sees a
token that looks like a decimal number with an invalid suffix, and it throws
an error.