Support ARM/IAR naming convention for wrapped functions
These linkers automatically perform wrapping on any symbols starting
$Sub$$ and $Super$$, as opposed to the GNU linker method, which uses
__wrap_ and __real_ prefixes and requires the use of command line switches
to activate them on a per-symbol basis.

When building assembly files in ArmDS, the 'as' tool has no in-built
predefines that we can use to distinguish it from 'as' assemblers in GNU
toolchains. So in order to convey the fact that we need to use the ARM
linker convention, add a definition in the toolchain file.

The prefixes use non-alphanumeric characters, so when used with the IAR
assembler will require the use of backtick quotes (and a mechanism to build
such names).