Don't add __wrap_ prefixes when building with IAR
The linker doesn't appear to provide the facility to convert references
from other compilation units to symbol using the same but with this magic
prefix. The absence of this just means we have to be more careful when
interworking the SDK with the system runtime library (i.e. IAR's runtime
library).