Have the IAR linker define symbols __StackTop and __StackOneTop
It is, however, unable to define __StackBottom or __StackOneBottom because
these depened upon the sizes of sections defined in object files. So rework
the C code that references these symbols to use __StackTop and __StackOneTop
when compiling with IAR.

These changes mean we now need to suppress diagnostic Pe170 which is
triggered by the pointer arithmetic on &__StackTop and &__StackOneTop.