Teach IAR EWARM to generate .pio.h files from .pio files using pioasm
This assumes a pre-built binary of the pioasm tool is present on PATH.
Packaging of the binaries is handled elsewhere.

We have to disable the usual rules to ensure that we don't attempt to
buidl host binaries (including picotool) when using this generator.

In order to have .pio files shown in the IAR IDE, they need to be added to
target_sources(). Custom command inputs aren't otherwise visible.