Teach IAR EWARM to generate .pio.h files from .pio files using pioasm
This requires a pre-built binary of the pioasm tool, since IAR is unable to
build its own host binaries. Add a GitHub Actions job to automatically build
a pioasm binary and raise a PR whenever it changes. To make detection of
binary changes reliable requires the use of the `ducible` tool to strip
timestamps and other nondeterministic data from the Windows PE executable
format.

Discussion is invited on the following topics:
* where do we want the .exe to live in the repository?
* how do we want the automatic PRs to be worded?
* how should the author field be set for the automatic PRs?