Rework handling of GCC's typeof operator
IAR (before version 9.30) assumes 'typeof' is an undeclared function symbol
and then generates a variety of parsing errors, including Pe109 'expression
preceding parentheses of apparent call must have (pointer-to-) function type',
Pe260 'explicit type is missing ("int" assumed)', Pe018 'expected a ")"' and
Pe075 'operand of "*" must be a pointer but has type "int"'.

An attempt had been made to support compilers that don't support typeof()
by testing __STRICT_ANSI__ but that symbol is itself a GCC extension and
isn't set by IAR (and can't really be be relied upon even in combination with
testing __GNUC__). So combine it with testing PICO_C_COMPILER_IS_IAR and
the (IAR-specific) __VER__.

For consistency, ensure we don't use inverted tests in some of the #if
statements, even if that means swapping over the if and else clauses.