Remove dependence on GCC's typeof operator
IAR (before version 9.30) assumes 'typeof' is an undeclared function symbol
and then generates a variety of parsing errors, including Pe109 'expression
preceding parentheses of apparent call must have (pointer-to-) function type',
Pe260 'explicit type is missing ("int" assumed)', Pe018 'expected a ")"' and
Pe075 'operand of "*" must be a pointer but has type "int"'.

There is no way to emulate the operator on earlier versions of IAR, so for
the two instances where macros are used which utilise typeof, provide an
alternative with hard-coded types.