Deal with inability of IAR linker to define useful symbols
Instead, for assembly, import start/end/length of linker blocks using the
block name suffixed by $$Base, $$Limit or $$Length (we can use the set
directive to assign local symbols equialent to what the GNU linker defines).
For C, use __section_begin(), __section_end() or __section_size() operators
(though those require the block name to have been predeclared as such using
a #pragma).

Under IAR, however, we are unable to reference an equivalent of
__StackBottom or __StackOneBottom directly, because these depened upon the
sizes of sections defined in object files. So rework the C code that
references these symbols to use __StackTop and __StackOneTop when
compiling with IAR.

These changes are responsible for the need to suppress diagnostic Pe170
which is triggered by the pointer arithmetic on &__StackTop and
&__StackOneTop.