Rework macro-local labels again
There are, in general, a numer of ways to express branches and labels in
macros that are invoked multiple times within the same source file, when
you can't use a normal label due to it being considered a redefinition of
a symbol:

* "altmacro" style local labels - not supported by armclang
* PC-relative expressions as branch targets - not appropriate in all cases,
  and also unsupported by armclang
* numeric labels - not supported by IAR
* macro counter ("\@") derived label names - not supported by IAR

As there is no option that works with all toolchains, introduce a couple
of C preprocessor macros to abstract the differences. These use the macro
counter for armclang and altmacro local labels for other toolchains. The
logic behind having GCC use local labels is that that scheme requires the
label names to be declared ahead of use, and since GCC builds are expected
to receive more regular testing, this will prevent build failures from
being introduced for IAR users.