Work around weak exported symbols not supported by armlink
This appears in practice to only be a problem for a couple of symbols, both
of which are multiply defined in both single- and double-precision FP
libraries. Armlink generates warning L6439W "Multiply defined Global Symbol
<symbol> defined in <section(object)> rejected in favor of Symbol defined
in <section(object)>."

Solved by making the definitions in the single-precision library dependent
upon the double-precsiion library not being included.