Introduce C preprocessor symbols for IAR and ARMClang
These follows the pattern of PICO_C_COMILER_IS_CLANG and
PICO_C_COMPILER_IS_GNU, and also corresponds to PICO_C_COMPILER_IS_IAR
and PICO_COMPILER_IS_ARMCLANG which are defined in the CMake scripts.

The first use of these desfines is to handle the fact that neither of
these toolchains ship with a <sys/cdefs.h> header file.