Fix type error with remove_volatile_cast
Used in newer versions of tinyusb, the previous implementation falls foul of
IAR's implementation of statement-expressions inheriting a void type from any
empty final subexpression. This typically results in error Pe144 'a value of
type "void" cannot be used to initialize an entity of [other] type'.

A possible alternative workaround would be to move the trailing semicolons
inside the definition of Clang_Pragma.